Вынесет часть логики из 'MovieCard' в 'Header'.
