+++ src/components/movie-card/movie-card.jsx
@@ -1,6 +1,7 @@
 import PropTypes from "prop-types";
 import React from "react";
-import {MoviePropType} from "../../prop-types";
+import {Header} from "../header/header.jsx";
+import {MoviePropType} from "../../prop-types.js";
 import {NUMBER_OF_ELEMENTS_IN_LINE, PageType} from "../../const.js";
 import {
   getRatingDescription,
@@ -11,11 +12,9 @@ import {
 
 export const MovieCard = (props) => {
   const {movie, activePage, onMovieClick} = props;
-  const {title,
-    backgroundUrl,
+  const {
+    title,
     posterUrl,
-    genres,
-    releaseDate,
     description,
     directors,
     starring,
@@ -31,14 +30,6 @@ export const MovieCard = (props) => {
     }
   };
 
-  const handleTitleClick = () => {
-    if (isMainIndexPage) {
-      onMovieClick();
-    }
-  };
-
-  const movieReleaseYear = releaseDate.getFullYear();
-  const genre = genres[0];
   const directorValue = getStringFromLimitedNumbersOfArrayElements(directors, NUMBER_OF_ELEMENTS_IN_LINE, `Director: `, ` and other`);
   const starringValue = getStringFromLimitedNumbersOfArrayElements(starring, NUMBER_OF_ELEMENTS_IN_LINE, `Starring: `, ` and other`);
   const scoreValue = getFormattedScore(score);
@@ -47,87 +38,13 @@ export const MovieCard = (props) => {
   const descriptionMarkup = description.map((descriptionItem, index) =>
     <p key={descriptionItem + index}>{descriptionItem}</p>);
 
-  const cardDeskMarkup = () => {
-    return (
-      <div className="movie-card__desc">
-        <h2
-          className="movie-card__title"
-          onClick={handleTitleClick}
-        >{title}</h2>
-        <p className="movie-card__meta">
-          <span className="movie-card__genre">{genre}</span>
-          <span className="movie-card__year">{movieReleaseYear}</span>
-        </p>
-
-        <div className="movie-card__buttons">
-          <button className="btn btn--play movie-card__button" type="button">
-            <svg viewBox="0 0 19 19" width="19" height="19">
-              <use xlinkHref="#play-s" />
-            </svg>
-            <span>Play</span>
-          </button>
-          <button className="btn btn--list movie-card__button" type="button">
-            <svg viewBox="0 0 19 20" width="19" height="20">
-              <use xlinkHref="#add" />
-            </svg>
-            <span>My list</span>
-          </button>
-          {isMainMovieDetailsPage && <a href="add-review.html" className="btn movie-card__button">Add review</a>}
-        </div>
-      </div>
-    );
-  };
-
-  const cardHeaderMarkup = () => {
-    const logoHref = isMainMovieDetailsPage ? {href: `main.html`} : null;
-
-    return (
-      <React.Fragment>
-        <div className="movie-card__bg">
-          <img src={backgroundUrl} alt={title} />
-        </div>
-
-        <h1 className="visually-hidden">WTW</h1>
-
-        <header className="page-header movie-card__head">
-          <div className="logo">
-            <a {...logoHref} className="logo__link">
-              <span className="logo__letter logo__letter--1">W</span>
-              <span className="logo__letter logo__letter--2">T</span>
-              <span className="logo__letter logo__letter--3">W</span>
-            </a>
-          </div>
-
-          <div className="user-block">
-            <div className="user-block__avatar">
-              <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
-            </div>
-          </div>
-        </header>
-
-        <div className="movie-card__wrap">
-          {isMainMovieDetailsPage && cardDeskMarkup() ||
-          <div className="movie-card__info">
-            <div className="movie-card__poster"
-              onClick={onMovieClick}
-            >
-              <img src={posterUrl} alt={title} width="218" height="327" />
-            </div>
-
-            {cardDeskMarkup()}
-          </div>}
-        </div>
-      </React.Fragment>
-    );
-  };
-
   return (
     <section className={`movie-card ${isMainMovieDetailsPage ? `movie-card--full` : ``}`}>
-      {
-        isMainMovieDetailsPage && <div className="movie-card__hero">
-          {cardHeaderMarkup()}
-        </div> || cardHeaderMarkup()
-      }
+      <Header
+        movie={movie}
+        activePage={activePage}
+        onMovieClick={onMovieClick}
+      />
 
       {isMainMovieDetailsPage &&
         <div className="movie-card__wrap movie-card__translate-top">
